fix: move tenants
